scraper/requirements.txt
selenium
supabase
python-dotenv
webdriver-manager
